Extract ConnectionId to darwin-types
