+++ darwin-types/Connection.ts
@@ -1,0 +1,1 @@
+export type ConnectionId = string;
