+++ darwin-types/messages/ConnectionInitialization.ts
@@ -1,5 +1,5 @@
 import { Message } from './Message';
-import { ConnectionId } from '../../game-backend/src/ServerStore';
+import { ConnectionId } from '../Connection';
 
 export interface ConnectionInitialization extends Message {
   type: 'connectionInitialization';
